Clean up some more things
