+++ .github/workflows/Test.yaml
@@ -16,11 +16,11 @@ env:
   AUTHENTICATION_ENABLED: false
   JWT_SECRET: Default
   JWT_EXPIRATION: 259200000
-  DATABASE_CLIENT: mysql2
+  DATABASE_CLIENT: pg
   DATABASE_HOST: 127.0.0.1
-  DATABASE_PORT: 3306
-  DATABASE_USER: root
-  DATABASE_PASSWORD: root
+  DATABASE_PORT: 5432
+  DATABASE_USER: postgres
+  DATABASE_PASSWORD: postgres
   INTERSERVICE_AUTHENTICATION_KEY: test
   BACKUP_RETENTION_DAYS: 30
 
@@ -37,10 +37,9 @@ jobs:
       - name: Check code style
         run: npm run prettier:check
 
-  # Depends on prettier finishing successfully
-  test-mysql:
+  test:
     needs: prettier
-    name: Run MySQL Tests (Node ${{ matrix.node-version }}) ${{ matrix.experimental && '(experimental)' || '' }}
+    name: Run Tests (Node ${{ matrix.node-version }}) ${{ matrix.experimental && '(experimental)' || '' }}
     runs-on: ubuntu-latest
     timeout-minutes: 5
     continue-on-error: ${{ matrix.experimental }}
@@ -52,14 +51,19 @@ jobs:
             experimental: false
           # - node-version: 22.x
           #   experimental: true
-
     services:
-      db:
-        image: mariadb:11.5.2-noble
+      postgres:
+        image: timescale/timescaledb:2.26.4-pg18
         env:
-          MARIADB_ROOT_PASSWORD: ${{ env.DATABASE_PASSWORD }}
+          POSTGRES_USER: postgres
+          POSTGRES_PASSWORD: postgres
         ports:
-          - 3306:3306
+          - 5432:5432
+        options: >-
+          --health-cmd "pg_isready -U postgres"
+          --health-interval 10s
+          --health-timeout 5s
+          --health-retries 5
     steps:
       - uses: actions/checkout@v6
       - name: Use Node.js ${{ matrix.node-version }}
@@ -72,75 +76,3 @@ jobs:
         run: npm test
       - name: Run API tests
         run: npm run test:api
-
-  test-postgres:
-    needs: prettier
-    name: Run PostgreSQL Tests
-    runs-on: ubuntu-latest
-    timeout-minutes: 5
-    env:
-      DATABASE_CLIENT: pg
-      DATABASE_PORT: 5432
-      DATABASE_USER: postgres
-      DATABASE_PASSWORD: postgres
-    services:
-      postgres:
-        image: postgres:16
-        env:
-          POSTGRES_PASSWORD: postgres
-        ports:
-          - 5432:5432
-        options: >-
-          --health-cmd "pg_isready -U postgres"
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-    steps:
-      - uses: actions/checkout@v6
-      - name: Use Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: ${{ env.NODE_VERSION }}
-      - name: Install dependencies
-        run: npm ci
-      - name: Run unit tests
-        run: npm test
-      - name: Run API tests
-        run: npm run test:api
-
-  test-bridge:
-    needs: prettier
-    name: Run Bridge Migration Tests
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    services:
-      db:
-        image: mariadb:11.5.2-noble
-        env:
-          MARIADB_ROOT_PASSWORD: ${{ env.DATABASE_PASSWORD }}
-        ports:
-          - 3306:3306
-      postgres:
-        image: postgres:16
-        env:
-          POSTGRES_PASSWORD: postgres
-        ports:
-          - 5432:5432
-        options: >-
-          --health-cmd "pg_isready -U postgres"
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-    steps:
-      - uses: actions/checkout@v6
-      - name: Use Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: ${{ env.NODE_VERSION }}
-      - name: Install dependencies
-        run: npm ci
-      - name: Run bridge migration tests
-        env:
-          BRIDGE_TEST_TARGET_USER: postgres
-          BRIDGE_TEST_TARGET_PASSWORD: postgres
-        run: npm run test:bridge --workspace=server
