+++ docker-compose.yaml
@@ -128,7 +128,7 @@ services:
       timeout: 3s
       retries: 20
     volumes:
-      - ./docker_volumes/postgres/data:/var/lib/postgresql/data
+      - ./docker_volumes/postgres:/var/lib/postgresql
     network_mode: "host"
     expose:
       - "5432"
